Day_6
오늘은 많이 진행하지 못했다.
Button을 상속해서 다양한 버튼을 구현할 예정이다.

+++ Assets/Scripts/Enemy.cs
@@ -31,7 +31,7 @@ public class Enemy : MonoBehaviour
         Debug.Log(m_HP);
     }
 
-    public void decreased(int power) //체력 감소
+    public void decreased(float power) //체력 감소
     {
         m_HP -= power;
         if (m_HP <= 0 && EnemyManager.isExist)
